requirements.txt
Flask
pandas
openpyxl
Pillow
Werkzeug
